refactor: move languages section to shared components
